buckle bucko we're in for the long haul
Loads textures from filepath. uses PIL to import the texture then converts it to the 6x6x6 colour encoding

the file directory is a bit of fucky thing. i dont understand it. when i run a python-3.9 and not vscode it copies the python files over to system 32. but not the texture files. so ive found a hacky way around that.

I think theres a bug with the texture filtering skipping a level of detail